requirements.txt
biopython>=1.83
rank_bm25>=0.1.2
sentence-transformers>=3.0.0
numpy>=1.24.0
groq>=0.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.0